fix: いいね機能APIの修正
